webapp-archive/requirements.txt
streamlit>=1.32.0
streamlit-authenticator>=0.3.1
streamlit-extras>=0.4.0
openai>=1.12.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
boto3>=1.34.0
stripe>=8.0.0
pillow>=10.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiohttp>=3.9.0
requests>=2.31.0
pydantic>=2.6.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.9